feat: created and used profile image input component
