+++ components/mentorDashboard/profile.js
@@ -1,10 +1,9 @@
 import React, { useEffect, useState } from "react";
 import axios from "axios";
-import Image from "next/image";
+import ProfileImageInput from "../basic/ProfileImageInput";
 import { FaUserAlt } from "react-icons/fa";
 import { BiSolidPhone, BiLogoLinkedin, BiLogoTwitter } from "react-icons/bi";
 import { MdEmail } from "react-icons/md";
-import { mentorImg } from "../../public/assets";
 import { ToastContainer, toast } from "react-toastify";
 
 function Profile({ mentorDetail, setLoadingState, setErrorState }) {
@@ -61,56 +60,12 @@ function Profile({ mentorDetail, setLoadingState, setErrorState }) {
     getMentorProfile();
   }, []);
 
-  // converting image to base64 function
-  const convertBase64 = (file) => {
-    return new Promise((resolve, reject) => {
-      const fileReader = new FileReader();
-      fileReader.readAsDataURL(file);
-
-      fileReader.onload = () => {
-        resolve(fileReader.result);
-      };
-
-      fileReader.onerror = (error) => {
-        reject(error);
-      };
-    });
-  };
-
   // onChange function for socials input
   const handleSocialChange = (e) => {
     setFormData({
       ...formData,
       social: { ...formData.social, [e.target.name]: e.target.value },
     });
-  };
-
-  const handleUploadImageChange = async (e) => {
-    const file = e.target.files[0];
-    const base64 = await convertBase64(file);
-
-    try {
-      const imageCloudinaryUrl = await uploadToCloudinary(base64);
-      console.log(imageCloudinaryUrl);
-      setFormData({ ...formData, image: imageCloudinaryUrl });
-    } catch (error) {
-      console.log("Error uploading image to Cloudinary:", error);
-    }
-  };
-
-  const uploadToCloudinary = async (imageSrc) => {
-    const res = await fetch(imageSrc);
-    const blob = await res.blob();
-    const url = `https://api.cloudinary.com/v1_1/${process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME}/image/upload`;
-    try {
-      const formData = new FormData();
-      formData.append("file", blob);
-      formData.append("upload_preset", "image_preset");
-      const res = await axios.post(url, formData);
-      return res.data.secure_url;
-    } catch (error) {
-      console.log("Couldn't upload image to Cloudinary", error);
-    }
   };
 
   // form submit function
@@ -152,20 +107,11 @@ function Profile({ mentorDetail, setLoadingState, setErrorState }) {
               className="mentorFormEdit max-[512px]:tw-justify-center max-[512px]:tw-items-center max-[512px]:tw-flex max-[512px]:tw-flex-col"
               onSubmit={handleSubmit}
             >
-              <div className="tw-mt-10 tw-items-center tw-flex tw-justify-center">
-                <Image
-                  className="tw-w-[100px] tw-h-[100px] tw-rounded-full tw-object-cover"
-                  src={formData.image ? formData.image : mentorImg}
-                  alt="mentor"
-                  width={100}
-                  height={100}
+                <ProfileImageInput
+                  image={formData.image}
+                  setImage={(newImage) => {setFormData({ ...formData, image: newImage })}}
+                  className="tw-mt-10 tw-items-center tw-flex tw-justify-center"
                 />
-                <input
-                  type="file"
-                  name="mentorProfile"
-                  onChange={(e) => handleUploadImageChange(e)}
-                />
-              </div>
               <div
                 style={{
                   display: "flex",
